login and logout

+++ src/utils/cloudinary.js
@@ -4,8 +4,13 @@ import fs from "fs" // inbuilt rehta hai
 cloudinary.config({
     cloud_name :process.env.CLOUDINARY_CLOUD_NAME,
     api_key : process.env.CLOUDINARY_API_KEY,
-    api_secret:process.env.CLOUDINARY_API_SECRECT,
+    api_secret:process.env.CLOUDINARY_API_SECRET,
 })
+
+console.log("Cloud:", process.env.CLOUDINARY_CLOUD_NAME)
+console.log("Key:", process.env.CLOUDINARY_API_KEY)
+console.log("Secret exists:", !!process.env.CLOUDINARY_API_SECRET)
+
 
 const uploadOnCloudinary = async (localFilePath) =>{
     try{
@@ -16,12 +21,20 @@ const uploadOnCloudinary = async (localFilePath) =>{
         })//.upload(filepath,options[])
         //file has been uploaded
         console.log(`successfully uploaded ${response.url}`); // .url gives public url
+        fs.unlinkSync(localFilePath) // local file ko delete krne ke liye use hota hai ,sync bcz hum chahte hai ki file delete ho jaye uske baad hi aage ka code execute ho
+        console.log(response);
         return response;
     }
     catch(error) {
-        fs.unlinkSync(localFilePath) // just remove the file from locally saved temp file as upload op failed
-        return null;
+    console.log("Cloudinary Error:");
+    console.log(error);
+
+    if(localFilePath){
+        fs.unlinkSync(localFilePath);
     }
+
+    return null;
+}
 }
 
 export {uploadOnCloudinary}
